service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
PyMuPDF==1.25.2
langchain==0.3.7
langchain-community==0.3.5
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.0
openai==1.54.0
anthropic==0.38.0
sentence-transformers==3.0.1
